services/users/db/database.py
import sqlalchemy as _sql
import sqlalchemy.ext.declarative as _declarative
import sqlalchemy.orm as _orm

URL = "postgresql://postgres:postgres@localhost/users"

engine = _sql.create_engine(URL)

sessionLocal = _orm.sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = _declarative.declarative_base()